# docs/swagger.yaml
definitions:
...
  models.AddressCreateI:
    properties:
      amphure:
        $ref: '#/definitions/models.BaseAmphureModel'
      name_en:
        type: string
      name_th:
        type: string
      post_code:
        type: integer
    type: object
  models.AddressDeleteI:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  models.AddressReadI:
    properties:
      filter:
        type: string
    type: object
  models.AddressReadO:
    properties:
      amphure:
        $ref: '#/definitions/models.BaseAmphureModel'
      id:
        type: string
      name_en:
        type: string
      name_th:
        type: string
      post_code:
        type: integer
    type: object
  models.AddressUpdateI:
    properties:
      amphure:
        $ref: '#/definitions/models.BaseAmphureModel'
      id:
        type: string
      name_en:
        type: string
      name_th:
        type: string
      post_code:
        type: integer
    type: object
  models.AddressUpdateO:
    properties:
      amphure:
        $ref: '#/definitions/models.BaseAmphureModel'
      id:
        type: string
      name_en:
        type: string
      name_th:
        type: string
      post_code:
        type: integer
    type: object
  models.BaseAmphureModel:
    properties:
      name_en:
        type: string
      name_th:
        type: string
      province:
        $ref: '#/definitions/models.BaseProvinceModel'
    type: object
  models.BaseProvinceModel:
    properties:
      name_en:
        type: string
      name_th:
        type: string
    type: object
  models.CreateProductGroupI:
    properties:
      groupName:
        type: string
    required:
    - groupName
    type: object
...
  models.CustomerCreateI:
    properties:
      address:
        type: string
      amphure:
        type: string
      contactNo:
        type: string
      email:
        type: string
      fullName:
        type: string
      postCode:
        type: integer
      province:
        type: string
      tambon:
        type: string
    required:
    - address
    - amphure
    - contactNo
    - email
    - fullName
    - postCode
    - province
    - tambon
    type: object
  models.CustomerDeleteI:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  models.CustomerReadI:
    properties:
      contactNo:
        type: string
      fullName:
        type: string
    type: object
  models.CustomerUpdateI:
    properties:
      address:
        type: string
      amphure:
        type: string
      contactNo:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      postCode:
        type: integer
      province:
        type: string
      tambon:
        type: string
    required:
    - id
    type: object
  models.DeleteProductGroupI:
    properties:
      id:
        type: string
    type: object
...
  models.ReadProductGroupO:
    properties:
      groupName:
        type: string
      id:
        type: string
    type: object
  models.ReadProductO:
    properties:
      barcodeId:
        type: string
      basePrice:
        type: number
      groupId:
        type: string
      id:
        type: string
      isBaseProduct:
        type: boolean
      productName:
        type: string
      productNameTh:
        type: string
      productSub:
        items:
          $ref: '#/definitions/models.ReadProductOSub'
        type: array
      salePrice:
        type: number
      unit:
        type: string
    type: object
  models.ReadProductOSub:
    properties:
      baseProductId:
        type: string
      baseProductUse:
        type: integer
      id:
        type: string
      productId:
        type: string
    type: object
...
  models.ResponseAPI-any:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  models.ResponseAPI-array_models_ReadProductGroupO:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.ReadProductGroupO'
        type: array
      message:
        type: string
    type: object
  models.ResponseAPI-array_models_ReadProductO:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.ReadProductO'
        type: array
      message:
        type: string
    type: object
...
  models.ResponseAPI-models_AddressReadO:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/models.AddressReadO'
      message:
        type: string
    type: object
  models.ResponseAPI-models_AddressUpdateO:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/models.AddressUpdateO'
      message:
        type: string
    type: object
...
  models.UpdateProductGroupI:
    properties:
      groupName:
        type: string
      id:
        type: string
    required:
    - id
    type: object
...
info:
  contact: {}
  description: This is a docs of Go Backend POS API
  title: Go Backend POS API
  version: "1.0"
paths:
  /api/v1/address:
    delete:
      description: to delete address
      operationId: Address
      parameters:
      - description: AddressDeleteI
        in: body
        name: AddressDeleteI
        schema:
          $ref: '#/definitions/models.AddressDeleteI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: Address
      tags:
      - Address
    get:
      description: to read address
      operationId: Address
      parameters:
      - description: AddressReadI
        in: body
        name: AddressReadI
        schema:
          $ref: '#/definitions/models.AddressReadI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-models_AddressReadO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: Address
      tags:
      - Address
    post:
      description: to create a new address
      operationId: Address
      parameters:
      - description: AddressCreateI
        in: body
        name: AddressCreateI
        schema:
          $ref: '#/definitions/models.AddressCreateI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: Address
      tags:
      - Address
    put:
      description: to update address
      operationId: Address
      parameters:
      - description: AddressUpdateI
        in: body
        name: AddressUpdateI
        schema:
          $ref: '#/definitions/models.AddressUpdateI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-models_AddressUpdateO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: Address
      tags:
      - Address
...
  /api/v1/customer:
    delete:
      description: to delete customer
      operationId: CustomerDelete
      parameters:
      - description: CustomerDeleteI
        in: body
        name: CustomerDeleteI
        schema:
          $ref: '#/definitions/models.CustomerDeleteI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: CustomerDelete
      tags:
      - Customer
    get:
      description: to read customer
      operationId: CustomerRead
      parameters:
      - description: CustomerReadI
        in: body
        name: CustomerReadI
        schema:
          $ref: '#/definitions/models.CustomerReadI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-array_models_ReadProductO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: CustomerRead
      tags:
      - Customer
    post:
      description: to create customer
      operationId: CustomerCreate
      parameters:
      - description: CustomerCreateI
        in: body
        name: CustomerCreateI
        schema:
          $ref: '#/definitions/models.CustomerCreateI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: CustomerCreate
      tags:
      - Customer
    put:
      description: to update customer
      operationId: CustomerUpdate
      parameters:
      - description: CustomerUpdateI
        in: body
        name: CustomerUpdateI
        schema:
          $ref: '#/definitions/models.CustomerUpdateI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: CustomerUpdate
      tags:
      - Customer
...
  /api/v1/productgroup:
    delete:
      description: to delete product group
      operationId: ProductGroupDelete
      parameters:
      - description: DeleteProductGroupI
        in: body
        name: DeleteProductGroupI
        schema:
          $ref: '#/definitions/models.DeleteProductGroupI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductGroupDelete
      tags:
      - ProductGroup
    get:
      description: to read product group
      operationId: ProductGroupRead
      parameters:
      - description: groupName
        in: query
        name: groupName
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-array_models_ReadProductGroupO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductGroupRead
      tags:
      - ProductGroup
    post:
      description: to create product group
      operationId: ProductGroupCreate
      parameters:
      - description: CreateProductGroupI
        in: body
        name: CreateProductGroupI
        schema:
          $ref: '#/definitions/models.CreateProductGroupI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductGroupCreate
      tags:
      - ProductGroup
    put:
      description: to update product group
      operationId: ProductGroupUpdate
      parameters:
      - description: UpdateProductGroupI
        in: body
        name: UpdateProductGroupI
        schema:
          $ref: '#/definitions/models.UpdateProductGroupI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductGroupUpdate
      tags:
      - ProductGroup
...
securityDefinitions:
  bearerAuth:
    description: 'Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"